validation/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
simplejson==3.19.2
pandas==1.5.3
fastapi==0.104.0
uvicorn==0.23.2
Jinja2==3.1.2
cvxpy==1.4.1
scikit-learn==1.3.2
scipy==1.11.3
python-multipart==0.0.6
great-expectations==0.18.0